third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/container-computed.html

<!DOCTYPE html>
<title>Computed values of container</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<div id="target"></div>
<script>
test_computed_value('container', 'initial', 'none');
test_computed_value('container', 'inherit', 'none');
test_computed_value('container', 'unset', 'none');
test_computed_value('container', 'inline-size');
test_computed_value('container', 'block-size');
test_computed_value('container', 'size');
test_computed_value('container', 'inline-size / inline-size');
test_computed_value('container', 'block-size / block-size');
test_computed_value('container', 'size / block-size');

test_computed_value('container', 'inline-size / foo');
test_computed_value('container', 'inline-size /foo', 'inline-size / foo');
test_computed_value('container', 'inline-size/ foo', 'inline-size / foo');
test_computed_value('container', 'inline-size/foo', 'inline-size / foo');
test_computed_value('container', 'size / FoO', 'size / FoO');
</script>
